Reorganise structs to own modules
The product state iterator and measurment enum have been placed in their
own file/module. No change has been made for interfacing with these
modules.
