+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "quantr"
-version = "0.5.2"
+version = "0.6.0"
 edition = "2021"
 license = "EUPL-1.2"
 readme = "README.md"
