+++ src/lib.rs
@@ -74,6 +74,6 @@ mod error;
 //  Make available for public use.
 pub use circuit::gate::Gate;
 pub use circuit::printer::Printer;
-pub use circuit::{states, Circuit, Measurement};
+pub use circuit::{measurement::Measurement, states, Circuit};
 pub use complex::{Complex, COMPLEX_ZERO};
 pub use error::QuantrError;
